sentinel_backend/requirements.txt
# Sentinel Backend - Python Dependencies
# =======================================
# Install with: pip install -r requirements.txt

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Browser Automation
playwright>=1.40.0

# WebSocket
websockets>=12.0

# PDF Generation (for audit reports)
fpdf>=1.7.2

# System Monitoring
psutil>=5.9.0

# Async utilities
aiohttp>=3.9.0
aiofiles>=23.2.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
